% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_infections_unit.R
\name{add_vaccination}
\alias{add_vaccination}
\title{add vaccination coverage to change immunity profile}
\usage{
add_vaccination(coverage, age_first, age_last, immunity, skew = 0)
}
\arguments{
\item{coverage}{vaccination campaign coverage}

\item{age_first}{youngest age group to be vaccinated}

\item{age_last}{oldest age group to be vaccinated}

\item{immunity}{current immunity in one year}

\item{skew}{skew of vaccination. Defaults to 0 where vaccination is random}
}
\value{
The immunity profile by age in one year
 @export
}
\description{
add vaccination coverage to change immunity profile
}
