% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cases_and_deaths.R
\name{calc_cases_and_deaths}
\alias{calc_cases_and_deaths}
\title{calculates the number of deaths and cases}
\usage{
calc_cases_and_deaths(new_infections, P_severe, P_severeDeath)
}
\arguments{
\item{new_infections}{number of infections}

\item{P_severe}{probability of severe infection}

\item{P_severeDeath}{probability of dying if severe}
}
\value{
the numbers of cases and deaths
}
\description{
calculates the number of deaths and cases
}
