% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DALYs.R
\name{calc_DALYs}
\alias{calc_DALYs}
\title{Calculate the number of DALYS}
\usage{
calc_DALYs(
  cases,
  deaths,
  life_exp_GAVI,
  P_severe,
  d_acute = 17.8/365,
  dw_acute = 0.172,
  d_conv = 28/365,
  dw_conv = 0.024
)
}
\arguments{
\item{cases}{number of cases per year, age and country}

\item{deaths}{number of deaths per year, age and country}

\item{life_exp_GAVI}{life expectancies from Montagu}

\item{P_severe}{probability of severe infection}

\item{d_acute}{days of acute infection per year. Default = 17.8 / 365.}

\item{dw_acute}{disability weight for acute. Default = 0.172.}

\item{d_conv}{days convalescent per year. Default = 28 / 365.}

\item{dw_conv}{disability weight if convalescent. Default = 0.024.}
}
\value{
DALYs by year, age and countries
}
\description{
Calculate the number of DALYS
}
