% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_infections.R
\name{calc_infections}
\alias{calc_infections}
\title{Calculate the number of infections}
\usage{
calc_infections(
  param_samples,
  vac_eff = 1,
  coverage_df,
  pop_all,
  years,
  age_max,
  model_type = "Foi"
)
}
\arguments{
\item{param_samples}{transmission intensity in each admin 1 unit or country}

\item{vac_eff}{vaccine efficacy, defaults to 1}

\item{coverage_df}{vaccination coverage}

\item{pop_all}{population in all countries by year and age}

\item{years}{years of interest}

\item{age_max}{maximum age in output so age range is [0, age_max]}

\item{model_type}{whether "Foi" or "R0". Defaults to "Foi".}
}
\value{
The number of infections in each country in each year of interest AND
        cohort size in each country in each year of interest AND immunity profiles.
}
\description{
Calculate the number of infections
}
