% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_vaccination.R
\name{compile_vaccination}
\alias{compile_vaccination}
\title{Compile vaccination coverages into one df}
\usage{
compile_vaccination(historic_dat, GAVI_preventive, WUENIC, GAVI_switch)
}
\arguments{
\item{historic_dat}{Historic vaccination coverages}

\item{GAVI_preventive}{GAVI preventive campaign info - this includes all vaccination activities}

\item{WUENIC}{WUENIC data for non-Gavi countries}

\item{GAVI_switch}{which vaccination activity to compile}
}
\value{
Vaccination activities for one GAVI switch scenario from 1940 to 2100
}
\description{
Compile vaccination coverages into one df
}
