% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Death_and_DALYs.R
\name{Death_and_DALYs}
\alias{Death_and_DALYs}
\title{calculates the number of deaths and DALYS}
\usage{
Death_and_DALYs(new_infections, P_severe, P_severeDeath, d_acute, dw_acute,
  d_conv, dw_conv, life_exp)
}
\arguments{
\item{new_infections}{number of infections}

\item{P_severe}{probability of severe infection}

\item{P_severeDeath}{probability od dying if severe}

\item{d_acute}{probability of acute infection?}

\item{dw_acute}{disability weight for acute}

\item{d_conv}{probability convalescent}

\item{dw_conv}{disability weight if convalescent}

\item{life_exp}{life expectancies}
}
\value{
the numbers of deaths and dalys in each country and year IN DEVELOPMENT
}
\description{
calculates the number of deaths and DALYS
}
