% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_immunityStart.R
\name{foi_prevac}
\alias{foi_prevac}
\title{calculate foi before vaccination starts for R0 model}
\usage{
foi_prevac(adm, R0, pop_moments, polydeg = 6)
}
\arguments{
\item{adm}{admin indeices}

\item{R0}{R0 in admins}

\item{pop_moments}{population moments}

\item{polydeg}{degree of polynomial. Defaults to 6}
}
\value{
The force of infection before vaccination in the admin or admins specified
}
\description{
calculate foi before vaccination starts for R0 model
}
