% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calcInfections.R
\name{fun_calcInfections}
\alias{fun_calcInfections}
\title{Calculate the number of infections}
\usage{
fun_calcInfections(param_samples, coverage_df, pop_all, years, age_max)
}
\arguments{
\item{param_samples}{transmission intensity in each admin 1 unit}

\item{coverage_df}{vaccination coverage}

\item{pop_all}{population in al countries by year and age}

\item{age_max}{maximum age in output so age range is [0, age_max]}

\item{year}{year of interest}
}
\value{
The number of infections in each country in each year of interest
fun_calcInfections()
}
\description{
Calculate the number of infections
}
