% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_immunityStart.R
\name{fun_immunityStart}
\alias{fun_immunityStart}
\title{calculate immunity profile at start}
\usage{
fun_immunityStart(model_type = "Foi", transmission_param, age_max, pop,
  old_coverage, year_end)
}
\arguments{
\item{model_type}{R0 or Foi. defaults to Foi}

\item{transmission_param}{transmission intensity in country}

\item{age_max}{maximum age gorup}

\item{pop}{population in country y year and age}

\item{old_coverage}{vaccination coverage by age and year}

\item{year_end}{year that we wish to have the immunity profile in}
}
\value{
immunity profile by age in the first year of interest
fun_immunityStart()
}
\description{
calculate immunity profile at start
}
