% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_infections_unit.R
\name{generate_infections_R0}
\alias{generate_infections_R0}
\title{calculate the number of infections in a year for all age groups in R0 model}
\usage{
generate_infections_R0(R0, pop, immunity)
}
\arguments{
\item{R0}{R0 in country}

\item{pop}{population by age in one year}

\item{immunity}{current immunity profile by age in one year}
}
\value{
The number of new infections in one year and the immunity profile by age in one year
}
\description{
calculate the number of infections in a year for all age groups in R0 model
}
