% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_infections_unit.R
\name{generate_infections_static}
\alias{generate_infections_static}
\title{calculate the number of infections in a year for all age groups in Foi model}
\usage{
generate_infections_static(foi, pop, immunity)
}
\arguments{
\item{foi}{Foi in country}

\item{pop}{population by age in one year}

\item{immunity}{current immunity profile by age in one year}
}
\value{
The immunity profile by age in one year and the number of infections for that year
generate_infections_static()
}
\description{
calculate the number of infections in a year for all age groups in Foi model
}
