% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DALYs.R
\name{reformat_life_expectancies}
\alias{reformat_life_expectancies}
\title{get life expectancies in array form}
\usage{
reformat_life_expectancies(life_exp_GAVI, years, ages)
}
\arguments{
\item{life_exp_GAVI}{life expectanices from Montagu}
}
\value{
life expectancies in form years x ages x countries
}
\description{
get life expectancies in array form
}
