% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_burden_for_template.R
\name{run_burden_for_template}
\alias{run_burden_for_template}
\title{run burden steps and put into a template}
\usage{
run_burden_for_template(historic_dat, GAVI_preventive, GAVI_switch,
  param_samples, pop_all, P_severe, P_severeDeath, life_exp_GAVI, template,
  run_id = NA)
}
\arguments{
\item{historic_dat}{= Historic vaccination coverages}

\item{GAVI_preventive}{GAVI preventive campaign info - this includes all vaccination activities}

\item{GAVI_switch}{which vaccination activity to compile}

\item{param_samples}{transmission intensity in each admin 1 unit or country}

\item{pop_all}{population in all countries by year and age}

\item{P_severe}{probability of severe infection}

\item{P_severeDeath}{probability of dying if severe}

\item{life_exp_GAVI}{life expectancies from Montagu}

\item{template}{output template}

\item{run_id}{run_id for stochastic runs. Defaults to NA}
}
\description{
run burden steps and put into a template
}
