% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_infections_unit.R
\name{update_immunity}
\alias{update_immunity}
\title{move on immunity profile through ageing}
\usage{
update_immunity(immunity)
}
\arguments{
\item{immunity}{current immunity profile}
}
\value{
New immunity profile by age
}
\description{
move on immunity profile through ageing
}
