% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_immunityStart.R
\name{fun_immunityStart}
\alias{fun_immunityStart}
\title{calculate immunity profile at start}
\usage{
fun_immunityStart(model_type = "Foi", transmission_param, age_max, pop,
  old_coverage)
}
\arguments{
\item{model_type}{R0 or Foi. defaults to Foi}

\item{transmission_param}{transmission intensity in country}

\item{age_max}{maximum age gorup}

\item{pop}{population in country by year and age}

\item{old_coverage}{vaccination coverage by year and age}
}
\value{
immunity profile by age in the first year of interest
}
\description{
calculate immunity profile at start
}
