% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_infections_unit.R
\name{run_infections_unit}
\alias{run_infections_unit}
\title{calculate the number of infections in a country for the years specified}
\usage{
run_infections_unit(model_type = "Foi", transmission_param,
  vac_eff = 1, years_in, age_max, pop, coverage, immunityStart)
}
\arguments{
\item{model_type}{whether R0 or Foi. Defaults to Foi}

\item{transmission_param}{transmission intensity in that country}

\item{vac_eff}{vaccine efficacy, defaults to 1}

\item{years_in}{years of interest}

\item{age_max}{maximum age group}

\item{pop}{population in countryby year and age}

\item{coverage}{vaccination coverage by campaign in country}

\item{immunityStart}{immunity profile at the beginning of simulation}
}
\value{
The immunity profile by age and number of infections in one year
}
\description{
calculate the number of infections in a country for the years specified
}
